requirements.txt
pydantic
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
defusedxml 
pypdf
bs4
arxiv
pymupdf
wikipedia
unstructured
faiss-cpu
langchain_google_genai
langchain_huggingface
langchain-pinecone
weaviate-client
langchain_weaviate
wikipedia
youtube-search
langgraph
chromadb
langchain-chroma